Add files via upload
The tasks has been named with respect to the order of the questions given. The orders were being sent using a seperate node created in order_sender.py and confirmations in the table and kitchen were sent using confirmation.py. The CancelOrder.srv file holds the service format which stores a boolean variable and checks if the order has been cancelled or not.